+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/verticalmode/SavedPaymentMethodRowButton.kt
@@ -84,7 +84,8 @@ internal fun SavedPaymentMethodRowButton(
         contentDescription = contentDescription,
         trailingContent = trailingContent,
         appearance = appearance,
-        shouldShowDefaultBadge = displayableSavedPaymentMethod.shouldShowDefaultBadge
+        shouldShowDefaultBadge = displayableSavedPaymentMethod.shouldShowDefaultBadge,
+        promotionProvider = null
     )
 }