+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/state/TapToAddConnectionStarterTest.kt
@@ -1,8 +1,11 @@
 package com.stripe.android.paymentsheet.state
 
 import com.google.common.truth.Truth.assertThat
+import com.stripe.android.common.model.CommonConfigurationFactory
 import com.stripe.android.common.taptoadd.FakeTapToAddConnectionManager
+import com.stripe.android.common.taptoadd.TapToAddConnectionManager
 import kotlinx.coroutines.test.StandardTestDispatcher
+import kotlinx.coroutines.test.advanceUntilIdle
 import kotlinx.coroutines.test.runTest
 import org.junit.Test
 
@@ -35,7 +38,7 @@ internal class TapToAddConnectionStarterTest {
     }
 
     @Test
-    fun `start calls manager connect`() = runTest(testDispatcher) {
+    fun `start calls manager connect with merchant display name from configuration`() = runTest(testDispatcher) {
         val manager = FakeTapToAddConnectionManager.noOp(isSupported = true)
         val starter = DefaultTapToAddConnectionStarter(
             tapToAddConnectionManager = manager,
@@ -43,10 +46,20 @@ internal class TapToAddConnectionStarterTest {
             coroutineContext = testDispatcher,
         )
 
-        starter.start()
-        testDispatcher.scheduler.advanceUntilIdle()
+        val commonConfiguration = CommonConfigurationFactory.create(
+            merchantDisplayName = "Books & Things",
+        )
 
-        assertThat(manager.connectCalls.awaitItem()).isNotNull()
+        starter.start(commonConfiguration)
+        advanceUntilIdle()
+
+        assertThat(manager.connectCalls.awaitItem()).isEqualTo(
+            FakeTapToAddConnectionManager.ConnectCall(
+                config = TapToAddConnectionManager.ConnectionConfig(
+                    merchantDisplayName = "Books & Things",
+                ),
+            )
+        )
         manager.connectCalls.ensureAllEventsConsumed()
     }
 }