+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/verticalmode/FakeSavedPaymentMethodConfirmInteractor.kt
@@ -5,24 +5,55 @@ import com.stripe.android.model.CardBrand
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.paymentsheet.DisplayableSavedPaymentMethod
 import com.stripe.android.paymentsheet.model.PaymentSelection
-import com.stripe.android.uicore.elements.FormElement
+import com.stripe.android.uicore.elements.IdentifierSpec
+import com.stripe.android.uicore.elements.SectionElement
+import com.stripe.android.uicore.elements.SimpleTextElement
+import com.stripe.android.uicore.elements.SimpleTextFieldConfig
+import com.stripe.android.uicore.elements.SimpleTextFieldController
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
 
-internal class FakeSavedPaymentMethodConfirmInteractor : SavedPaymentMethodConfirmInteractor {
-    override val displayableSavedPaymentMethod: DisplayableSavedPaymentMethod = DisplayableSavedPaymentMethod.create(
-        displayName = "···· 4242".resolvableString,
-        paymentMethod = PaymentMethod(
-            id = "pm_123",
-            created = null,
-            liveMode = false,
-            code = PaymentMethod.Type.Card.code,
-            type = PaymentMethod.Type.Card,
-            card = PaymentMethod.Card(
-                brand = CardBrand.Visa,
-                last4 = "4242",
-            )
+internal class FakeSavedPaymentMethodConfirmInteractor(
+    formEnabled: Boolean = false,
+) : SavedPaymentMethodConfirmInteractor {
+    private val _state = MutableStateFlow(
+        SavedPaymentMethodConfirmInteractor.State(
+            displayableSavedPaymentMethod = DisplayableSavedPaymentMethod.create(
+                displayName = "···· 4242".resolvableString,
+                paymentMethod = PaymentMethod(
+                    id = "pm_123",
+                    created = null,
+                    liveMode = false,
+                    code = PaymentMethod.Type.Card.code,
+                    type = PaymentMethod.Type.Card,
+                    card = PaymentMethod.Card(
+                        brand = CardBrand.Visa,
+                        last4 = "4242",
+                    )
+                ),
+            ),
+            form = SavedPaymentMethodConfirmInteractor.State.Form(
+                elements = listOf(
+                    SectionElement.wrap(
+                        sectionFieldElements = listOf(
+                            SimpleTextElement(
+                                identifier = IdentifierSpec.Generic("Name"),
+                                controller = SimpleTextFieldController(
+                                    initialValue = "John Doe",
+                                    textFieldConfig = SimpleTextFieldConfig(
+                                        label = "Name".resolvableString
+                                    )
+                                )
+                            )
+                        )
+                    )
+                ),
+                enabled = formEnabled,
+            ),
         )
     )
-    override val formElement: FormElement? = null
+    override val state: StateFlow<SavedPaymentMethodConfirmInteractor.State> = _state.asStateFlow()
 
     class Factory : SavedPaymentMethodConfirmInteractor.Factory {
         override fun create(