+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # CHANGELOG
 
-NEXT_VERSION_BUMP: PATCH
+NEXT_VERSION_BUMP: MINOR
 
 ## XX.XX.XX - 20XX-XX-XX
+* [REMOVED][12871](https://github.com/stripe/stripe-android/pull/12871) Removed UPI support across the SDK.
+
+### PaymentSheet
+* [ADDED][12880](https://github.com/stripe/stripe-android/pull/12880) Added support for Pay by Bank (GA in GB, private preview in EU).
 
 ## 23.4.0 - 2026-04-14
 ### Crypto Onramp (Private Preview)